+++ FrontEnd/src/app/layouts/admin-layout/rrhh/rrhh.component.ts
@@ -83,16 +83,28 @@ export class RrhhComponent implements OnInit {
   }
 
   deleteEmployee(_id: string) {
-    this._rS.deleteEmployee(_id)
-      .subscribe(res => {
-        this.getRrhh().then(() => {
-          Swal.fire(
-            'Ok!',
-            'Empleado eliminado correctamente',
-            'success'
-          ).then(() => this.loading = false)
-        });
-      });
+    Swal.fire({
+      title: 'Desea eliminar el empleado?',
+      text: "No podrá revertir este proceso",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._rS.deleteEmployee(_id)
+          .subscribe(res => {
+            this.getRrhh().then(() => {
+              Swal.fire(
+                'Ok!',
+                'Empleado eliminado correctamente',
+                'success'
+              ).then(() => this.loading = false)
+            });
+          });
+      }
+    })
   }
 
 }
